Changed the storing arrays of byte to byte arrays instead of short arrays. Refactored some fields. Removed some useless classes. Stops the player is the newly opened song has an unknown format. Added a "Go to music folder" icon in the PlayMusic Activity.

+++ src/aks/jnv/reader/ISongReader.java
@@ -54,7 +54,7 @@ public interface ISongReader {
 	 * Returns the registers of the next frame of the song.
 	 * @return the registers of the next frame of the song.
 	 */
-	short[] getNextRegisters();
+	int[] getNextRegisters();
 	
 	/**
 	 * Seeks to the given second in the song.
@@ -81,7 +81,7 @@ public interface ISongReader {
 	 * @param sampleNumber the zero-based index of the sample.
 	 * @return the sample, or Null.
 	 */
-	short[] getSample(int sampleNumber);
+	byte[] getSample(int sampleNumber);
 	
 	
 	/**
